requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
tornado==6.1
tzdata==2021.1
tzlocal==3.0
psycopg2==2.9.1
asgiref==3.4.1
charset-normalizer==2.0.7
idna==3.3
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1